+++ custom_components/bambu_lab/switch.py
@@ -5,7 +5,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import EntityCategory
 
-from .const import DOMAIN, LOGGER
+from .const import DOMAIN, LOGGER, Options
 from .models import BambuLabEntity
 from .pybambu.const import Features
 
@@ -59,6 +59,7 @@ TIMELAPSE_SWITCH_DESCRIPTION = SwitchEntityDescription(
     entity_category=EntityCategory.CONFIG,
 )
 
+
 async def async_setup_entry(
         hass: HomeAssistant,
         entry: ConfigEntry,
@@ -84,6 +85,7 @@ async def async_setup_entry(
 
     if coordinator.get_model().supports_feature(Features.TIMELAPSE):
         async_add_entities([BambuLabTimelapseSwitch(coordinator, entry)])
+
 
 class BambuLabSwitch(BambuLabEntity, SwitchEntity):
     """Base BambuLab Switch"""
@@ -112,7 +114,7 @@ class BambuLabManualModeSwitch(BambuLabSwitch):
             config_entry: ConfigEntry
     ) -> None:
         super().__init__(coordinator, config_entry)
-        self._attr_is_on = self.coordinator.client.manual_refresh_mode
+        self._attr_is_on = self.coordinator.get_option_enabled(Options.MANUALREFRESH)
 
     @property
     def available(self) -> bool:
@@ -125,13 +127,14 @@ class BambuLabManualModeSwitch(BambuLabSwitch):
 
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Enable manual refresh mode."""
-        self._attr_is_on = not self.coordinator.client.manual_refresh_mode
-        await self.coordinator.set_manual_refresh_mode(True)
+        self._attr_is_on = True
+        await self.coordinator.set_option_enabled(Options.MANUALREFRESH, True)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Disable manual refresh mode."""
-        self._attr_is_on = not self.coordinator.client.manual_refresh_mode
-        await self.coordinator.set_manual_refresh_mode(False)
+        self._attr_is_on = False
+        await self.coordinator.set_option_enabled(Options.MANUALREFRESH, False)
+
 
 class BambuLabCameraSwitch(BambuLabSwitch):
     """BambuLab Refresh data Switch"""
@@ -144,7 +147,7 @@ class BambuLabCameraSwitch(BambuLabSwitch):
             config_entry: ConfigEntry
     ) -> None:
         super().__init__(coordinator, config_entry)
-        self._attr_is_on = self.coordinator.camera_enabled
+        self._attr_is_on = self.coordinator.get_option_enabled(Options.CAMERA)
 
     @property
     def available(self) -> bool:
@@ -158,12 +161,12 @@ class BambuLabCameraSwitch(BambuLabSwitch):
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Enable the camera."""
         self._attr_is_on = True
-        await self.coordinator.set_camera_enabled(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.CAMERA, True)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Disable the camera."""
         self._attr_is_on = False
-        await self.coordinator.set_camera_enabled(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.CAMERA, False)
 
 
 class BambuLabCameraImageSwitch(BambuLabSwitch):
@@ -177,11 +180,11 @@ class BambuLabCameraImageSwitch(BambuLabSwitch):
             config_entry: ConfigEntry
     ) -> None:
         super().__init__(coordinator, config_entry)
-        self._attr_is_on = self.coordinator.camera_as_image_sensor
+        self._attr_is_on = self.coordinator.get_option_enabled(Options.IMAGECAMERA)
 
     @property
     def available(self) -> bool:
-        return self.coordinator.camera_enabled
+        return self.coordinator.get_option_enabled(Options.CAMERA)
 
     @property
     def icon(self) -> str:
@@ -191,12 +194,12 @@ class BambuLabCameraImageSwitch(BambuLabSwitch):
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Enable the camera."""
         self._attr_is_on = True
-        await self.coordinator.set_camera_as_image_sensor(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.IMAGECAMERA, self._attr_is_on)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Disable the camera."""
         self._attr_is_on = False
-        await self.coordinator.set_camera_as_image_sensor(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.IMAGECAMERA, self._attr_is_on)
 
 
 class BambuLabFtpSwitch(BambuLabSwitch):
@@ -210,7 +213,7 @@ class BambuLabFtpSwitch(BambuLabSwitch):
             config_entry: ConfigEntry
     ) -> None:
         super().__init__(coordinator, config_entry)
-        self._attr_is_on = self.coordinator.ftp_enabled
+        self._attr_is_on = self.coordinator.get_option_enabled(Options.FTP)
 
     @property
     def available(self) -> bool:
@@ -224,12 +227,12 @@ class BambuLabFtpSwitch(BambuLabSwitch):
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Enable FTP."""
         self._attr_is_on = True
-        await self.coordinator.set_ftp_enabled(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.FTP, self._attr_is_on)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Disable FTP."""
         self._attr_is_on = False
-        await self.coordinator.set_ftp_enabled(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.FTP, self._attr_is_on)
 
 
 class BambuLabTimelapseSwitch(BambuLabSwitch):
@@ -243,7 +246,7 @@ class BambuLabTimelapseSwitch(BambuLabSwitch):
             config_entry: ConfigEntry
     ) -> None:
         super().__init__(coordinator, config_entry)
-        self._attr_is_on = self.coordinator.timelapse_enabled
+        self._attr_is_on = self.coordinator.get_option_enabled(Options.TIMELAPSE)
 
     @property
     def available(self) -> bool:
@@ -257,12 +260,12 @@ class BambuLabTimelapseSwitch(BambuLabSwitch):
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Enable Timelapse Download."""
         self._attr_is_on = True
-        await self.coordinator.set_timelapse_enabled(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.TIMELAPSE, self._attr_is_on)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Disable Timelapse Download."""
         self._attr_is_on = False
-        await self.coordinator.set_timelapse_enabled(self._attr_is_on)
+        await self.coordinator.set_option_enabled(Options.TIMELAPSE, self._attr_is_on)
 
 
 class BambuLabPromptSoundSwitch(BambuLabSwitch):