+++ frontend/src/const.ts
@@ -1,0 +1,3 @@
+export const PREFIX_NAME = "ha-bambulab";
+export const INTEGRATION_DOMAIN = "bambu_lab";
+export const MANUFACTURER = "Bambu Lab";