+++ frontend/src/ha-bambulab-cards.ts
@@ -1,0 +1,4 @@
+import { version } from "../package.json";
+import "./cards/ams-card/ams-card";
+
+console.info(`%c🐼 Bambu Lab 🐼 - ${version}`, "color: #ef5350; font-weight: 700;");